+++ lib/domain/api_client/movie_and_tv_api_client.dart
@@ -6,79 +6,8 @@ import 'package:comics_db_app/domain/entity/popular_tv_response.dart';
 import 'package:comics_db_app/domain/entity/trending_all_response.dart';
 import 'package:comics_db_app/domain/entity/tv_details.dart';
 
-/*
-1. нет сети
-2. нет ответа, таймаут соединения
-
-3. сервер недоступен
-4. сервер не может обработать запрашиваемый запрос
-5. сервер ответил не то, что мы ожидали
-
-6. сервер ответил ожидаемой ошибкой
-*/
-
-enum MediaType { movie, tv }
-
-extension MediaTypeAsString on MediaType {
-  String asString() {
-    switch (this) {
-      case MediaType.movie:
-        return 'movie';
-      case MediaType.tv:
-        return 'tv';
-    }
-  }
-}
-
-class ApiClient {
+class MovieAndTvApiClient {
   final _networkClient = NetworkClient();
-
-  Future<int> getAccountInfo(String sessionId) async {
-    int parser(dynamic json) {
-      final jsonMap = json as Map<String, dynamic>;
-      final result = jsonMap['id'] as int;
-      return result;
-    }
-
-    final result = _networkClient.get(
-      '/account',
-      parser,
-      <String, dynamic>{
-        'api_key': Configuration.apiKey,
-        'session_id': sessionId,
-        // 'movieId': movieId.toString(),
-      },
-    );
-    return result;
-  }
-
-  Future<int> markAsFavorite({
-    required int accountId,
-    required String sessionId,
-    required MediaType mediaType,
-    required int mediaId,
-    required bool isFavorite,
-  }) async {
-    int parser(dynamic json) {
-      return 1;
-    }
-
-    final parameters = <String, dynamic>{
-      'media_type': mediaType.asString(),
-      'media_id': mediaId,
-      'favorite': isFavorite,
-    };
-    final result = _networkClient.post(
-      '/account/$accountId/favorite',
-      parameters,
-      parser,
-      <String, dynamic>{
-        'api_key': Configuration.apiKey,
-        'session_id': sessionId
-      },
-    );
-    return result;
-  }
 
   Future<PopularAndTopRatedMovieResponse> popularMovie(
       int page, String locale) async {
@@ -361,7 +290,6 @@ class ApiClient {
     );
     return result;
   }
-
 }
 
 /*
