+++ lib/ui/widgets/movie_top_rated/top_rated_movie_model.dart
@@ -2,7 +2,7 @@
 
 import 'dart:async';
 
-import 'package:comics_db_app/domain/api_client/api_client.dart';
+import 'package:comics_db_app/domain/api_client/movie_and_tv_api_client.dart';
 import 'package:comics_db_app/domain/entity/movie.dart';
 import 'package:comics_db_app/domain/entity/popular_and_top_rated_movie_response.dart';
 import 'package:comics_db_app/ui/navigation/main_navigation.dart';
@@ -10,7 +10,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:intl/intl.dart';
 
 class TopRatedMovieModel extends ChangeNotifier {
-  final _apiClient = ApiClient();
+  final _apiClient = MovieAndTvApiClient();
   final _movies = <Movie>[];
   late int _currentPage;
   late int _totalPage;
