+++ lib/ui/widgets/trending/trending_all_model.dart
@@ -1,11 +1,11 @@
-import 'package:comics_db_app/domain/api_client/api_client.dart';
+import 'package:comics_db_app/domain/api_client/movie_and_tv_api_client.dart';
 import 'package:comics_db_app/domain/entity/trending_all.dart';
 import 'package:comics_db_app/domain/entity/trending_all_response.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:intl/intl.dart';
 
 class TrendingAllModel extends ChangeNotifier {
-  final _apiClient = ApiClient();
+  final _apiClient = MovieAndTvApiClient();
   final _trendingAll = <TrendingAll>[];
   late int _currentPage;
   late int _totalPage;
