+++ lib/ui/widgets/tv_details/tv_details_model.dart
@@ -1,4 +1,5 @@
-import 'package:comics_db_app/domain/api_client/api_client.dart';
+import 'package:comics_db_app/domain/api_client/account_api_client.dart';
+import 'package:comics_db_app/domain/api_client/movie_and_tv_api_client.dart';
 import 'package:comics_db_app/domain/api_client/api_client_exception.dart';
 import 'package:comics_db_app/domain/data_providers/session_data_provider.dart';
 import 'package:comics_db_app/domain/entity/tv_details.dart';
@@ -7,7 +8,8 @@ import 'package:intl/intl.dart';
 
 class TvDetailsModel extends ChangeNotifier {
   final _sessionDataProvider = SessionDataProvider();
-  final _apiClient = ApiClient();
+  final _movieAndTvApiClient = MovieAndTvApiClient();
+  final _accountApiClient = AccountApiClient();
   final int tvId;
   bool _isFavoriteTV = false;
   String _locale = '';
@@ -32,10 +34,10 @@ class TvDetailsModel extends ChangeNotifier {
 
   Future<void> loadTVDetails() async {
     try {
-      _tvDetails = await _apiClient.tvDetails(tvId, _locale);
+      _tvDetails = await _movieAndTvApiClient.tvDetails(tvId, _locale);
       final sessionId = await _sessionDataProvider.getSessionId();
       if (sessionId != null) {
-        _isFavoriteTV = await _apiClient.isFavoriteTV(tvId, sessionId);
+        _isFavoriteTV = await _movieAndTvApiClient.isFavoriteTV(tvId, sessionId);
       }
       notifyListeners();
     } on ApiClientException catch (e) {
@@ -52,7 +54,7 @@ class TvDetailsModel extends ChangeNotifier {
     _isFavoriteTV = !_isFavoriteTV;
     notifyListeners();
     try {
-      await _apiClient.markAsFavorite(
+      await _accountApiClient.markAsFavorite(
           accountId: accountId,
           sessionId: sessionId,
           mediaType: MediaType.tv,
